# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kiosco Franco API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: string;
  name: string;
  price: number;
  category: string;
  emoji: string;
  image?: string | null;
}

export interface ProductInput {
  name: string;
  price: number;
  category: string;
  emoji: string;
  image?: string | null;
}

export interface OrderItem {
  productId: string;
  name: string;
  price: number;
  qty: number;
}

export type OrderDelivery = (typeof OrderDelivery)[keyof typeof OrderDelivery];

export const OrderDelivery = {
  retiro: "retiro",
  envio: "envio",
} as const;

export type OrderPayment = (typeof OrderPayment)[keyof typeof OrderPayment];

export const OrderPayment = {
  efectivo: "efectivo",
  mercadopago: "mercadopago",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  nuevo: "nuevo",
  preparacion: "preparacion",
  listo: "listo",
  entregado: "entregado",
} as const;

export interface Order {
  id: string;
  orderNumber?: number;
  createdAt: string;
  customerName: string;
  address: string;
  delivery: OrderDelivery;
  payment: OrderPayment;
  items: OrderItem[];
  total: number;
  status: OrderStatus;
}

export type OrderInputDelivery =
  (typeof OrderInputDelivery)[keyof typeof OrderInputDelivery];

export const OrderInputDelivery = {
  retiro: "retiro",
  envio: "envio",
} as const;

export type OrderInputPayment =
  (typeof OrderInputPayment)[keyof typeof OrderInputPayment];

export const OrderInputPayment = {
  efectivo: "efectivo",
  mercadopago: "mercadopago",
} as const;

export interface OrderInput {
  customerName: string;
  address: string;
  delivery: OrderInputDelivery;
  payment: OrderInputPayment;
  items: OrderItem[];
}

export type OrderStatusUpdateStatus =
  (typeof OrderStatusUpdateStatus)[keyof typeof OrderStatusUpdateStatus];

export const OrderStatusUpdateStatus = {
  nuevo: "nuevo",
  preparacion: "preparacion",
  listo: "listo",
  entregado: "entregado",
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface Settings {
  shopName: string;
  whatsappNumber: string;
  mercadoPagoAlias: string;
}