# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kiosco Franco API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: products
  - name: orders
  - name: settings
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [orders]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Settings"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: integer
        category:
          type: string
        emoji:
          type: string
        image:
          type: string
          nullable: true
      required: [id, name, price, category, emoji]

    ProductInput:
      type: object
      properties:
        name:
          type: string
        price:
          type: integer
        category:
          type: string
        emoji:
          type: string
        image:
          type: string
          nullable: true
      required: [name, price, category, emoji]

    OrderItem:
      type: object
      properties:
        productId:
          type: string
        name:
          type: string
        price:
          type: integer
        qty:
          type: integer
      required: [productId, name, price, qty]

    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: integer
        createdAt:
          type: string
          format: date-time
        customerName:
          type: string
        address:
          type: string
        delivery:
          type: string
          enum: [retiro, envio]
        payment:
          type: string
          enum: [efectivo, mercadopago]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: integer
        status:
          type: string
          enum: [nuevo, preparacion, listo, entregado]
      required:
        [id, createdAt, customerName, address, delivery, payment, items, total, status]

    OrderInput:
      type: object
      properties:
        customerName:
          type: string
        address:
          type: string
        delivery:
          type: string
          enum: [retiro, envio]
        payment:
          type: string
          enum: [efectivo, mercadopago]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required: [customerName, address, delivery, payment, items]

    OrderStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [nuevo, preparacion, listo, entregado]
      required: [status]

    Settings:
      type: object
      properties:
        shopName:
          type: string
        whatsappNumber:
          type: string
        mercadoPagoAlias:
          type: string
        mercadoPagoQr:
          type: string
          nullable: true
      required: [shopName, whatsappNumber, mercadoPagoAlias]