# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kiosco Franco API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderDelivery } from "./orderDelivery";
import type { OrderItem } from "./orderItem";
import type { OrderPayment } from "./orderPayment";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  orderNumber?: number;
  createdAt: Date;
  customerName: string;
  address: string;
  delivery: OrderDelivery;
  payment: OrderPayment;
  items: OrderItem[];
  total: number;
  status: OrderStatus;
}