# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kiosco Franco API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputDelivery } from "./orderInputDelivery";
import type { OrderInputPayment } from "./orderInputPayment";
import type { OrderItem } from "./orderItem";

export interface OrderInput {
  customerName: string;
  address: string;
  delivery: OrderInputDelivery;
  payment: OrderInputPayment;
  items: OrderItem[];
}