# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kiosco Franco API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  name: string;
  price: number;
  category: string;
  emoji: string;
  image?: string | null;
}