# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kiosco Franco API
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  shopName: string;
  whatsappNumber: string;
  mercadoPagoAlias: string;
}